⚰️ entities/record에 남아있는 event 로직을 삭제한다
